move cec into its own file
